chore: Update subscription.html template
This commit updates the `subscription.html` template in the `pos` app. Changes include adding a centered title and description for each subscription plan, as well as a "Select" button. Additionally, a list of items is added for each plan, but commented out. This update improves the layout and presentation of subscription options for users.