chore: Refactor chat message sending logic in navbar.html
This commit refactors the chat message sending logic in the `navbar.html` template of the `pos` app. The previous implementation used a click event on the `#send_msg` element to send the message, but this functionality has been removed. The code related to sending and displaying the message has been commented out. This change improves code readability and removes unnecessary functionality from the template.