feat: Update Sale model date_added field to use auto_now_add
This commit updates the `date_added` field in the `Sale` model of the `sales` app to use the `auto_now_add` option. This change ensures that the `date_added` field is automatically set to the current date and time when a new sale is created. By using `auto_now_add`, we eliminate the need to manually set the `date_added` field and improve the accuracy of the sale timestamps.