feat: Update sidenav.html to fix broken links
This commit fixes the broken links in the sidenav.html template of the pos app. The href attributes for the sidebar links were empty, causing the links to lead to incorrect destinations. By updating the href attributes to the correct URLs, we ensure that users can navigate to the intended sections of the application through the sidebar menu.