```text refactor: Update sales_this_week view to fix date range bug
The commit updates the sales_this_week view in the reports app to fix a bug in the date range. Previously, the date range was set from the 1st to the 7th day of the current month, but it should be from the 1st to the 8th day. This change ensures that the sales for the current week are accurately retrieved.