```text feat: Remove taxtype_id and alter taxTypeCode field in TaxType model
This commit removes the taxtype_id field and alters the taxTypeCode field in the TaxType model in the etims app. The taxtype_id field is no longer needed, and the alteration of the taxTypeCode field enhances the functionality of the application by making it the primary key and ensuring its uniqueness.