refactor: Update base.html for improved script and stylesheet loading
This commit updates the base.html template to improve the loading of scripts and stylesheets. It removes the unnecessary select2 and select2-bootstrap4-theme CSS files and replaces the Bootstrap Touchspin script and stylesheet with the latest versions. Additionally, it updates the jQuery and Bootstrap scripts to their latest versions. These changes optimize the loading process and ensure compatibility with the latest dependencies.