feat: Add purchase details page template
A new HTML template has been added for displaying the details of a purchase order. This template includes the purchase ID, as well as a card displaying the purchase information. The template also includes buttons for navigating back to the purchases index page and editing the purchase. This addition improves the user experience by providing a dedicated page for viewing and managing individual purchase orders.