feat: Create new expense category form template
A new HTML template has been added for creating a new expense category. This template includes form fields for entering the category name and description, as well as a checkbox for indicating if the category is recurring. The addition of this template improves the user experience by providing a dedicated page for creating new expense categories.