feat: Add barcode field to Product model
This commit adds a new barcode field to the Product model in the Django POS application. The barcode field is a CharField with a maximum length of 20 characters. It is nullable and can be left blank. This addition allows for the storage and retrieval of barcode information for each product, enhancing the functionality of the application.