feat: Add "Add Branch" functionality to Django POS
This commit adds the "Add Branch" functionality to the Django POS application. It includes the necessary changes to the `urls.py` file to add a new URL path for adding a branch. This addition enhances the functionality of the application by allowing users to add new branches through the user interface.